chore: refactor folder structure for heroku
