+++ server/db/config/config.js
@@ -1,17 +1,18 @@
-import dotenv from 'dotenv';
+import dotenv from "dotenv";
 
 dotenv.config();
 
-module.exports = {
+const config = {
   development: {
     url: process.env.DATABASE_URL_DEV,
-    dialect: 'postgres'
+    dialect: "postgres",
   },
   test: {
-    url: process.env.DATABASE_URL_TEST,
-    dialect: 'postgres'
+    dialect: "postgres",
   },
   production: {
-    dialect: 'postgres'
-  }
+    dialect: "postgres",
+  },
 };
+
+export default config;
