+++ server/routes/index.js
@@ -1,0 +1,5 @@
+const router = require("express").Router();
+
+router.use("/api", require("./api"));
+
+module.exports = router;
